fixes/refactor post update merge, moved consolidation to consolidator
